+++ src/main/java/clothe_management/controller/entity/ReceivingDetail.java
@@ -4,24 +4,24 @@ public class ReceivingDetail {
     private String receivingID;
     private String productID;
     private int quantity;
-    private double unitCostPrice;
-    public ReceivingDetail(String receivingID, String productID, int quantity, double unitCostPrice) {
+    private int unitCostPrice;
+    public ReceivingDetail(String receivingID, String productID, int quantity, int unitCostPrice) {
         this.receivingID = receivingID;
         this.productID = productID;
         this.quantity = quantity;
         this.unitCostPrice = unitCostPrice;
     }
-    public String getReceivingID() { 
-        return receivingID; 
+    public String getReceivingID() {
+        return receivingID;
     }
-    public String getProductID() { 
-        return productID; 
+    public String getProductID() {
+        return productID;
     }
-    public int getQuantity() { 
-        return quantity; 
+    public int getQuantity() {
+        return quantity;
     }
-    public double getUnitCostPrice() {
-        return unitCostPrice; 
+    public int getUnitCostPrice() {
+        return unitCostPrice;
     }
     public void setReceivingID(String receivingID) {
         this.receivingID = receivingID;
@@ -32,7 +32,7 @@ public class ReceivingDetail {
     public void setQuantity(int quantity) {
         this.quantity = quantity;
     }
-    public void setUnitCostPrice(double unitCostPrice) {
+    public void setUnitCostPrice(int unitCostPrice) {
         this.unitCostPrice = unitCostPrice;
     }
 }
